Introduce typed supervised desktop runtime
