+++ .github/workflows/release-draft.yml
@@ -93,7 +93,7 @@ jobs:
           "INNO_COMPILER=$compiler" >> $env:GITHUB_ENV
 
       - name: Compile
-        run: python -m py_compile Windows/main.py Windows/amazon_devtools.py Windows/media_reader.py Windows/discord_rpc.py Windows/config.py Windows/credential_store.py Windows/network_audit.py Windows/updater.py Windows/status_summary.py Windows/qt_tray_ui.py Windows/rpc_state.py Windows/metadata_pipeline.py Windows/launcher_diagnostics.py Windows/security_trust.py Windows/self_tests.py Windows/release_smoke.py Windows/run_self_tests.py
+        run: python -m py_compile Windows/main.py Windows/desktop_runtime.py Windows/app_models.py Windows/runtime_state.py Windows/task_supervisor.py Windows/structured_logging.py Windows/window_controller.py Windows/rpc_controller.py Windows/command_controller.py Windows/amazon_devtools.py Windows/media_reader.py Windows/discord_rpc.py Windows/config.py Windows/credential_store.py Windows/network_audit.py Windows/updater.py Windows/status_summary.py Windows/qt_tray_ui.py Windows/rpc_state.py Windows/metadata_pipeline.py Windows/launcher_diagnostics.py Windows/security_trust.py Windows/self_tests.py Windows/release_smoke.py Windows/run_self_tests.py
 
       - name: Pytest
         run: python -m pytest Windows/tests
@@ -103,6 +103,9 @@ jobs:
 
       - name: Ruff safety checks
         run: python -m ruff check Windows --select E9,F63,F7,F82
+
+      - name: Typed controller checks
+        run: python -m pyright Windows/app_models.py Windows/runtime_state.py Windows/task_supervisor.py Windows/structured_logging.py Windows/window_controller.py Windows/rpc_controller.py Windows/command_controller.py Windows/main.py
 
       - name: Dependency audit
         run: python -m pip_audit -r Windows/requirements-release.lock --strict
