+++ Windows/AmazonMusicRPC.spec
@@ -26,6 +26,14 @@ a = Analysis(
         'metadata_pipeline',
         'launcher_diagnostics',
         'security_trust',
+        'desktop_runtime',
+        'app_models',
+        'runtime_state',
+        'task_supervisor',
+        'structured_logging',
+        'window_controller',
+        'rpc_controller',
+        'command_controller',
         'credential_store',
         'network_audit',
         'qt_tray_ui',
