+++ Windows/amazify_rpc_bridge.py
@@ -1,9 +1,11 @@
 import json
+import hmac
+import re
 import threading
 import time
 from http.server import BaseHTTPRequestHandler, ThreadingHTTPServer
 
-from amazify_compat import AMAZIFY_RPC_BRIDGE_PORT
+from amazify_compat import AMAZIFY_RPC_BRIDGE_PORT, amazify_bridge_token
 from config import APP_VERSION
 
 
@@ -17,16 +19,20 @@ ALLOWED_COMMANDS = {
     "toggle_rpc",
     "updates",
 }
+AMAZON_ORIGIN_RE = re.compile(r"^https://(?:music|www)\.amazon\.(?:com|de|co\.uk|fr|it|es|co\.jp|ca|com\.au|com\.br|com\.mx)$", re.IGNORECASE)
+TOKEN_HEADER = "X-Amazon-Music-RPC-Token"
+MAX_REQUEST_BYTES = 16 * 1024
 
 
 class AmazifyRpcBridge:
-    def __init__(self, state_provider, command_handler, port=AMAZIFY_RPC_BRIDGE_PORT):
+    def __init__(self, state_provider, command_handler, port=AMAZIFY_RPC_BRIDGE_PORT, token=None):
         self.state_provider = state_provider
         self.command_handler = command_handler
         self.port = int(port)
         self.httpd = None
         self.thread = None
         self.started_at = time.time()
+        self.token = str(token or amazify_bridge_token())
 
     def start(self):
         if self.httpd:
@@ -38,9 +44,15 @@ class AmazifyRpcBridge:
                 return
 
             def do_OPTIONS(self):
+                if not self.allowed_origin:
+                    self._send_json({"ok": False, "error": "Origin not allowed"}, status=403)
+                    return
                 self._send_json({"ok": True})
 
             def do_GET(self):
+                if not self.authorized:
+                    self._send_json({"ok": False, "error": "Unauthorized"}, status=401)
+                    return
                 if self.path_only == "/health":
                     self._send_json({"ok": True, "app": "AmazonMusicRPC", "version": APP_VERSION, "port": bridge.port})
                     return
@@ -55,6 +67,9 @@ class AmazifyRpcBridge:
                 self._send_json({"ok": False, "error": "Not found"}, status=404)
 
             def do_POST(self):
+                if not self.authorized:
+                    self._send_json({"ok": False, "error": "Unauthorized"}, status=401)
+                    return
                 if self.path_only != "/command":
                     self._send_json({"ok": False, "error": "Not found"}, status=404)
                     return
@@ -78,8 +93,23 @@ class AmazifyRpcBridge:
             def path_only(self):
                 return self.path.split("?", 1)[0]
 
+            @property
+            def allowed_origin(self):
+                origin = str(self.headers.get("Origin") or "")
+                return origin if AMAZON_ORIGIN_RE.fullmatch(origin) else ""
+
+            @property
+            def authorized(self):
+                supplied = str(self.headers.get(TOKEN_HEADER) or "")
+                return bool(supplied and hmac.compare_digest(supplied, bridge.token))
+
             def _read_payload(self):
-                length = int(self.headers.get("Content-Length") or "0")
+                try:
+                    length = int(self.headers.get("Content-Length") or "0")
+                except ValueError as error:
+                    raise ValueError("Invalid Content-Length") from error
+                if length < 0 or length > MAX_REQUEST_BYTES:
+                    raise ValueError("Request body is too large")
                 raw = self.rfile.read(length) if length else b"{}"
                 try:
                     data = json.loads(raw.decode("utf-8") or "{}")
@@ -94,9 +124,11 @@ class AmazifyRpcBridge:
                 self.send_response(status)
                 self.send_header("Content-Type", "application/json; charset=utf-8")
                 self.send_header("Content-Length", str(len(body)))
-                self.send_header("Access-Control-Allow-Origin", "*")
-                self.send_header("Access-Control-Allow-Methods", "GET, POST, OPTIONS")
-                self.send_header("Access-Control-Allow-Headers", "Content-Type")
+                if self.allowed_origin:
+                    self.send_header("Access-Control-Allow-Origin", self.allowed_origin)
+                    self.send_header("Vary", "Origin")
+                    self.send_header("Access-Control-Allow-Methods", "GET, POST, OPTIONS")
+                    self.send_header("Access-Control-Allow-Headers", f"Content-Type, {TOKEN_HEADER}")
                 self.send_header("Cache-Control", "no-store")
                 self.end_headers()
                 self.wfile.write(body)
@@ -105,7 +137,7 @@ class AmazifyRpcBridge:
             self.httpd = ThreadingHTTPServer(("127.0.0.1", self.port), Handler)
         except OSError:
             return False
-        self.thread = threading.Thread(target=self.httpd.serve_forever, name="amazify-rpc-bridge", daemon=True)
+        self.thread = threading.Thread(target=self.httpd.serve_forever, name="amazify-rpc-bridge", daemon=False)
         self.thread.start()
         return True
 
@@ -118,6 +150,9 @@ class AmazifyRpcBridge:
         except Exception:
             pass
         self.httpd = None
+        if self.thread and self.thread is not threading.current_thread():
+            self.thread.join(timeout=5)
+        self.thread = None
 
 
 def start_amazify_rpc_bridge(state_provider, command_handler):
