+++ Windows/listenbrainz_scrobbler.py
@@ -1,9 +1,10 @@
 # MIT License - Copyright (c) 2026 eripum9
 
-import time
 import traceback
 import threading
 import liblistenbrainz
+
+from task_supervisor import TaskSupervisor
 
 
 class ListenBrainzScrobbler:
@@ -12,6 +13,7 @@ class ListenBrainzScrobbler:
         self.client.set_auth_token(user_token, check_validity=False)
         self._pending = []
         self._lock = threading.Lock()
+        self._tasks = TaskSupervisor()
 
     def update_now_playing(self, title, artist, album=None, duration=None):
         self._run_async("now playing", self._update_now_playing_sync, title, artist, album, duration, drop_if_busy=True)
@@ -81,4 +83,8 @@ class ListenBrainzScrobbler:
             finally:
                 self._lock.release()
 
-        threading.Thread(target=runner, daemon=True).start()
+        self._tasks.start_unique(f"listenbrainz-{label}", runner)
+
+    def close(self):
+        self._tasks.request_stop()
+        self._tasks.join(timeout=3)
